Move settings from installations to preferences
