+++ UnitystationLauncher/ViewModels/PreferencesPanelViewModel.cs
@@ -1,11 +1,15 @@
-﻿using System.Threading.Tasks;
+﻿using System;
+using System.Threading.Tasks;
+using System.Reactive.Linq;
 using MsBox.Avalonia.Base;
 using ReactiveUI;
 using Serilog;
 using UnitystationLauncher.Constants;
 using UnitystationLauncher.Infrastructure;
-using UnitystationLauncher.Models.Enums;
 using UnitystationLauncher.Services.Interface;
+using UnitystationLauncher.Models.Enums;
+using UnitystationLauncher.Models.ConfigFile;
+using DynamicData.Binding;
 
 namespace UnitystationLauncher.ViewModels
 {
@@ -21,17 +25,104 @@ namespace UnitystationLauncher.ViewModels
             get => _installationPath;
             set => this.RaiseAndSetIfChanged(ref _installationPath, value);
         }
+        private bool _autoRemove;
+        public bool AutoRemove
+        {
+            get => _autoRemove;
+            set => this.RaiseAndSetIfChanged(ref _autoRemove, value);
+        }
+		// for a future overhaul of TTS
+        // private string TTSPath;
+        // public string TTSText
+        // { // exp ? true : false
+        //     get { return string.IsNullOrEmpty(TTSPath) ? "Not Installed." : $"Is Installed at {TTSPath}"; }
+        // }
+        // public string TTSButtonText
+        // {
+        //     get { return string.IsNullOrEmpty(TTSPath) ? "Install" : "Uninstall"; }
+        // }
+        private bool? _TTSEnabled;
+
+        public bool? TTSEnabled
+        {
+            get => _TTSEnabled;
+            set => this.RaiseAndSetIfChanged(ref _TTSEnabled, value);
+        }
 
         private readonly IPreferencesService _preferencesService;
         private readonly IInstallationService _installationService;
+		private readonly IEnvironmentService _environmentService;
+		private readonly ITTSService _ttsService;
 
-        public PreferencesPanelViewModel(IPreferencesService preferencesService, IInstallationService installationService)
+        public PreferencesPanelViewModel(
+			IPreferencesService preferencesService, 
+			IInstallationService installationService,
+			IEnvironmentService environmentService,
+			ITTSService ttsService)
         {
             _preferencesService = preferencesService;
             _installationService = installationService;
+			_environmentService = environmentService;
+			_ttsService = ttsService;
 
-            _installationPath = _preferencesService.GetPreferences().InstallationPath;
+			Preferences preferences = _preferencesService.GetPreferences();
+            _installationPath = preferences.InstallationPath;
+			_autoRemove = preferences.AutoRemove;
+			_TTSEnabled = preferences.TTSEnabled;
+			this.WhenAnyValue(p => p.AutoRemove)
+                .Select(_ => Observable.FromAsync(OnAutoRemoveChangedAsync))
+                .Concat()
+                .Subscribe();
+
+            this.WhenAnyValue(p => p.TTSEnabled)
+                .Select(_ => Observable.FromAsync(OnTTSChangedAsync))
+                .Concat()
+                .Subscribe();
         }
+
+		public async Task OnAutoRemoveChangedAsync() {
+			if (AutoRemove)
+            {
+                IMsBox<string> msgBox = MessageBoxBuilder.CreateMessageBox(MessageBoxButtons.YesNo,
+                    "Are you sure?", "This will remove older installations from disk. Proceed?");
+
+                string response = await msgBox.ShowAsync();
+				AutoRemove = response.Equals(MessageBoxResults.Yes);
+            }
+			_preferencesService.GetPreferences().AutoRemove = AutoRemove;
+		}
+
+		public async Task OnTTSChangedAsync() {
+			if (_environmentService.GetCurrentEnvironment() == CurrentEnvironment.MacOsStandalone)
+            {
+                IMsBox<string> msgBox = MessageBoxBuilder.CreateMessageBox(MessageBoxButtons.Ok,
+                    " MacOS is unsupported ",
+                    " Text-To-Speech is not supported on MacOS. If you would like to contribute, talk to us on the discord. ");
+                await msgBox.ShowAsync(); //should this even be awaited?
+                return;
+            }
+
+
+            if (TTSEnabled is false)
+            {
+                IMsBox<string> msgBox = MessageBoxBuilder.CreateMessageBox(MessageBoxButtons.YesNo,
+                    "Are you sure?", "This will disable Character voices and delete the TTS System. This may take a While.");
+                string response = await msgBox.ShowAsync();
+				TTSEnabled = response.Equals(MessageBoxResults.No);
+            }
+        	_preferencesService.GetPreferences().TTSEnabled = TTSEnabled;
+
+            if (TTSEnabled == false)
+            {
+                _ttsService.StopTTS();
+                string installationBasePath = _preferencesService.GetPreferences().InstallationPath; //yucky
+                var LocalVersion = System.IO.Path.Combine(installationBasePath, "tts");
+                if (System.IO.Directory.Exists(LocalVersion))
+                {
+                    System.IO.Directory.Delete(LocalVersion, true);
+                }
+            }
+		}
 
         public async Task SetInstallationPathAsync(string path)
         {
